fix(erros): middleware global de exceção + Sum(decimal) quebrado no SQLite local
Não existia nenhum tratamento global de exceção não tratada — um erro que
escapasse de um controller virava 500 vazio (produção) ou página de erro em
HTML do .NET (dev), sem JSON nenhum pro frontend mostrar. Explica o "some
sem mensagem" que o QA reportou: o toast genérico do frontend não tinha
nada real pra exibir.

Novo middleware (primeiro do pipeline, envolve tudo): loga a exceção
completa no servidor e devolve JSON consistente ({ message, traceId }) —
mensagem real em Development, mensagem genérica + traceId (pra cruzar
com o log do servidor) em produção.

Testado ao vivo contra uma loja vazia local: revelou um bug de verdade —
Sum() em campo decimal não traduz pro SQL no SQLite (dev local sem
Postgres), só o Postgres de produção sabe. FinanceiroCalculoService e
GeminiChatService corrigidos pro mesmo padrão já usado em outro ponto do
próprio FinanceiroCalculoService: soma como (long), divide por 100m depois
em memória — sem perda de precisão, funciona nos dois bancos.

183/183 testes passando.

+++ CardGameStore/Services/Implementations/FinanceiroCalculoService.cs
@@ -51,8 +51,11 @@ public class FinanceiroCalculoService : IFinanceiroCalculoService
                 c.SecondPaymentMethod == filterPaymentMethod);
 
         // ── Receita de comandas ───────────────────────────────────────────────
+        // Sum() traduzido pro SQL precisa ser (long), não (decimal) — SQLite (dev
+        // local sem Postgres) não sabe agregar decimal no banco, só Postgres sabe.
+        // Divide por 100m depois, em memória, sem perder precisão.
         var receitaComandas = await comandasBaseQ
-            .SumAsync(c => (decimal)c.TotalInCents) / 100m;
+            .SumAsync(c => (long)c.TotalInCents) / 100m;
 
         // ── Vendas avulsas ──────────────────────────────────────────
         var todasVendas = (await _vendas.GetRecentAsync(2000, ini)).ToList();
@@ -106,9 +109,10 @@ public class FinanceiroCalculoService : IFinanceiroCalculoService
         var margemPercent = custo > 0 ? Math.Round(margem / custo * 100, 1) : 0;
 
         // ── Crediários em aberto ──────────────────────────────────────────────
+        // Mesmo motivo do Sum de receita acima: (long), não (decimal), pra traduzir no SQLite.
         var crediarios = await _db.Crediarios
             .Where(c => c.Status == CrediariosStatus.Aberto)
-            .SumAsync(c => (decimal)(c.ValorEmCentavos - c.ValorPagoEmCentavos)) / 100m;
+            .SumAsync(c => (long)(c.ValorEmCentavos - c.ValorPagoEmCentavos)) / 100m;
 
         // ── Breakdown dia a dia ───────────────────────────────────────────────
         // Uma passada só por cada lista, acumulando num dicionário por dia BR
